Stutter issue fixed
Got rid of the use of java.swing.Timer objects in the Model and for each Animation which fixed the lag. They now use the system time to determine when events need to happen. 
A core program loop has been added to the Driver to regulate the speed at which the whole program updates.